+++ seagulls-engine/src/seagulls/rendering/__init__.py
@@ -1,13 +1,13 @@
 from ._color import Color, ColorDict
 from ._game_screen import IGameScreen, IProvideGameScreens
 from ._position import Position, PositionDict
-from ._printer import IPrintSquares
+from ._printer import IPrinter
 from ._size import Size, SizeDict
 
 __all__ = [
     "IGameScreen",
     "IProvideGameScreens",
-    "IPrintSquares",
+    "IPrinter",
     "ColorDict",
     "Color",
     "PositionDict",
