+++ seagulls-engine/src/seagulls/rendering/_camera.py
@@ -1,9 +1,10 @@
 import logging
+from pathlib import Path
 from typing import Tuple
 
 from seagulls.rendering import (
     Color,
-    IPrintSquares,
+    IPrinter,
     Position,
     Size
 )
@@ -12,7 +13,7 @@ from seagulls.rendering._position import IUpdatePosition
 logger = logging.getLogger(__name__)
 
 
-class Camera(IPrintSquares, IUpdatePosition):
+class Camera(IPrinter, IUpdatePosition):
 
     """
     2022-04-16
@@ -24,67 +25,85 @@ class Camera(IPrintSquares, IUpdatePosition):
     - end the frame?
     """
 
-    _square_printer: IPrintSquares
+    _printer: IPrinter
     _size: Size
     _position: Position
     _background_color: Color
 
     def __init__(
             self,
-            square_printer: IPrintSquares,
+            printer: IPrinter,
             size: Size,
             position: Position):
-        self._square_printer = square_printer
+        self._printer = printer
         self._size = size
         self._position = position
 
         self._background_color = Color({"r": 0, "g": 0, "b": 0})
 
+    def print_sprite(self, file: Path, size: Size, position: Position) -> None:
+        object_position = position.get()
+        camera_position = self._position.get()
+
+        try:
+            self._assert_objects_overlap(size, position)
+            adjusted_position = Position({
+                "x": object_position["x"] - camera_position["x"],
+                "y": object_position["y"] - camera_position["y"],
+            })
+            self._printer.print_sprite(file, size, adjusted_position)
+        except ObjectDoesNotOverlapError:
+            logger.warning("Skipping!")
+            pass
+
     def print_square(self, color: Color, size: Size, position: Position):
+        object_position = position.get()
+        camera_position = self._position.get()
+
+        try:
+            self._assert_objects_overlap(size, position)
+            adjusted_position = Position({
+                "x": object_position["x"] - camera_position["x"],
+                "y": object_position["y"] - camera_position["y"],
+            })
+            self._printer.print_square(color, size, adjusted_position)
+        except ObjectDoesNotOverlapError:
+            logger.warning("Skipping!")
+            pass
+
+    def _assert_objects_overlap(self, size: Size, position: Position) -> None:
         object_position = position.get()
         object_size = size.get()
 
         camera_position = self._position.get()
         camera_size = self._size.get()
 
-        def _assert_objects_overlap() -> None:
-            # This sounds like something we could do with collision detection
-            if object_position["x"] + object_size["width"] < camera_position["x"]:
-                # Object is too far to the left to be visible
-                logger.warning("Object is too far to the left to be visible")
-                raise ObjectDoesNotOverlapError()
+        # This sounds like something we could do with collision detection
+        if object_position["x"] + object_size["width"] < camera_position["x"]:
+            # Object is too far to the left to be visible
+            logger.warning("Object is too far to the left to be visible")
+            raise ObjectDoesNotOverlapError()
 
-            if object_position["y"] + object_size["height"] < camera_position["y"]:
-                # Object is too far above to be visible
-                logger.warning("Object is too far above to be visible")
-                raise ObjectDoesNotOverlapError()
+        if object_position["y"] + object_size["height"] < camera_position["y"]:
+            # Object is too far above to be visible
+            logger.warning("Object is too far above to be visible")
+            raise ObjectDoesNotOverlapError()
 
-            if object_position["x"] > camera_position["x"] + camera_size["width"]:
-                # Object is too far to the right to be visible
-                logger.warning("# Object is too far to the right to be visible")
-                raise ObjectDoesNotOverlapError()
+        if object_position["x"] > camera_position["x"] + camera_size["width"]:
+            # Object is too far to the right to be visible
+            logger.warning("# Object is too far to the right to be visible")
+            raise ObjectDoesNotOverlapError()
 
-            if object_position["y"] > camera_position["y"] + camera_size["height"]:
-                # Object is too far below to be visible
-                logger.warning("Object is too far below to be visible")
-                raise ObjectDoesNotOverlapError()
-
-        try:
-            _assert_objects_overlap()
-            adjusted_position = Position({
-                "x": object_position["x"] - camera_position["x"],
-                "y": object_position["y"] - camera_position["y"],
-            })
-            self._square_printer.print_square(color, size, adjusted_position)
-        except ObjectDoesNotOverlapError:
-            logger.warning("Skipping!")
-            pass
+        if object_position["y"] > camera_position["y"] + camera_size["height"]:
+            # Object is too far below to be visible
+            logger.warning("Object is too far below to be visible")
+            raise ObjectDoesNotOverlapError()
 
     def commit(self) -> None:
-        self._square_printer.commit()
+        self._printer.commit()
 
     def clear(self):
-        self._square_printer.clear()
+        self._printer.clear()
 
     def update_position(self, position: Position) -> None:
         self._position = position
