+++ seagulls-engine/src/seagulls/rendering/_example/_renderables.py
@@ -1,0 +1,72 @@
+import random
+from pathlib import Path
+from typing import Tuple
+
+from seagulls.rendering import IPrinter, SizeDict, Color, Size, Position
+from seagulls.rendering._example._components import SolidColorComponent, SpriteComponent
+from seagulls.rendering._renderable_component import IProvideRenderables, RenderableComponent
+
+
+class MyRenderables(IProvideRenderables):
+
+    _printer: IPrinter
+    _scene_size: SizeDict
+
+    def __init__(
+            self,
+            printer: IPrinter,
+            scene_size: SizeDict):
+        self._printer = printer
+        self._scene_size = scene_size
+
+    def get(self) -> Tuple[RenderableComponent, ...]:
+        color = Color({
+            "r": random.randint(0, 255),
+            "g": random.randint(0, 255),
+            "b": random.randint(0, 255),
+        })
+        size = Size({
+            "height": random.randint(1, 3),
+            "width": random.randint(1, 3),
+        })
+        position1 = Position({
+            "x": random.randint(0, 50),
+            "y": random.randint(0, 50),
+        })
+        position2 = Position({
+            "x": random.randint(self._scene_size["width"] - 50,
+                                self._scene_size["width"]),
+
+            "y": random.randint(self._scene_size["height"] - 50,
+                                self._scene_size["height"]),
+        })
+        return tuple([
+            SolidColorComponent(
+                color=color,
+                size=size,
+                position=position1,
+                printer=self._printer,
+            ),
+            SolidColorComponent(
+                color=color,
+                size=size,
+                position=position2,
+                printer=self._printer,
+            ),
+            SpriteComponent(
+                filepath=Path(
+                    "seagulls-rpg-demo/seagulls_assets/sprites/environment/"
+                    "rpg-environment-options/tile_0060.png"),
+                size=Size({"height": 16, "width": 16}),
+                position=Position({"x": 200, "y": 50}),
+                printer=self._printer,
+            ),
+            SpriteComponent(
+                filepath=Path(
+                    "seagulls-rpg-demo/seagulls_assets/sprites/environment/"
+                    "rpg-environment-options/tile_0026.png"),
+                size=Size({"height": 16*4, "width": 16*4}),
+                position=Position({"x": 100, "y": 50}),
+                printer=self._printer,
+            ),
+        ])
